requirements.txt
# Core dependencies
click
python-dotenv
notion-client
sentence-transformers
langchain
langchain-community
langchain-core
langchain-experimental
chromadb

# Transformers / Embeddings 

torch
transformers
numpy
Pillow

# Utils
requests

# Groq client
groq
